third_party/blink/web_tests/external/wpt/websockets/back-forward-cache-with-open-websocket-connection.window-expected.txt

This is a testharness.js-based test.
[FAIL] Testing BFCache support for page with open WebSocket connection.
  assert_equals: document unexpectedly BFCached expected (undefined) undefined but got (boolean) true
Harness: the test ran to completion.
